[clang] Move IdentifierLoc to IdentifierTable.h and fix assertion build errors
