+++ clang/include/clang/Basic/SourceLocation.h
@@ -467,29 +467,6 @@ public:
   }
 };
 
-/// A simple pair of identifier info and location.
-class IdentifierLoc {
-  SourceLocation Loc;
-  IdentifierInfo *II = nullptr;
-
-public:
-  IdentifierLoc() = default;
-  IdentifierLoc(SourceLocation L, IdentifierInfo *Ident) : Loc(L), II(Ident) {}
-
-  void setLoc(SourceLocation L) { Loc = L; }
-  void setIdentifierInfo(IdentifierInfo *Ident) { II = Ident; }
-  SourceLocation getLoc() const { return Loc; }
-  IdentifierInfo *getIdentifierInfo() const { return II; }
-
-  bool operator==(const IdentifierLoc &X) const {
-    return Loc == X.Loc && II == X.II;
-  }
-
-  bool operator!=(const IdentifierLoc &X) const {
-    return Loc != X.Loc || II != X.II;
-  }
-};
-
 } // namespace clang
 
 namespace llvm {
